[PDK-14675][Sciclient]:Fix incomplete execution logs for sciclient fw testapp
Currently, a loop variable is getting incremented infinite times and hence a
for loop in sciclient_fw_testapp is running in an infinite loop due to
which the execution logs are incomplete.

This could be due to compiler optimization on the loop variable and hence
changed the loop variable to volatile inorder to fix this problem.

Signed-off-by: Thejasvi Konduru <t-konduru@ti.com>