[PDK-13573][Sciclient]:Fix the description for the calculation of MSMC Cache size
The cache is 4-way set associative and the value of n determines the number of sets
in the cache. Each cache line stores (2^15) Bytes. For example, if n=8 then the cache
size will be (n*4*(2^15)/(2^20)) MB = n/8 MB which is n*X/64 where X is the total size
of the MSMC memory which is 8MB in j784s4.

The comment in the sciclient_defaultBoardcfg.c is fixed accordingly based on the
above analysis.

Signed-off-by: Thejasvi Konduru <t-konduru@ti.com>