UDMA_UT: Fix channel,rm negative test cases on j7200
Issue-1:
The coverage of the negative rm test cases on j7200 is not achieved as the functions are returning failure even before the lines expected.
This is because of the zero start and number of resources allocated to the respective Rx/Tx channels. It is fixed by configuring the parameters with non zero values.
Updated UdmaRmAllocTxChTestNeg, UdmaRmTestNeg, UdmaRmAllocRxChTestNeg and UdmaRmAllocBlkCopyChTestNeg test cases with non-zero parameters.
Issue-2:
CUrrently PDMA and PSIL tests are covered with UTC_PRESENT flag and j7200 doesnot have UTCs, therefore these testcases are not getting executed in j7200.
PDMA and PSIL are not dependent on UTCs, therefore moving these tests out of UTC_PRESENT flag in UdmaTestChOpenNeg.
Issue-3
"else" branch in Tx/Rx chPause and chResume are covered in all other SOCs apart from j7200 by UTC tests.
Updated UdmaChDisableTest to cover Tx/Rx chPause and chResume APIs in j7200.

Signed-off-by: Sushmi x1123751@ti.com