[PDK-13985][SCICLIENT]:Fix sciclient_firmware_boot_testapp for hs and hsfs devices
Sciclient_firmware_boot_testapp is using the SCICLIENT_FIRMWARE macro of GP device even
for HS and HSFS devices.

This is fixed by allowing the user to specify the type of device during the build of this
testcase. The makefile corresponding to the testcase is modified inorder to differentiate
between GP, HS and HSFS devices. The name of the testapp should be as follows for different
type of devices.

GP   : sciclient_firmware_boot_testapp_gp
HS   : sciclient_firmware_boot_testapp_hs
HS-FS: sciclient_firmware_boot_testapp_hsfs

Signed-off-by: Thejasvi Konduru <t-konduru@ti.com>