[Sciclient]:Update calls to Hwip_disable() and Hwip_restore()
When a thread executing Sciclient_osalAcquireSecureProxyAcess() can reach a break statement
depending on the arguments passed, so in that case there is currently a mismatch in the no.of
times HwiP_disable() and HwiP_restore() are called.

Hence the code is modified inorder to have an equal no.of calls to HwiP_disable() and Hwip_restore().

Signed-off-by: Thejasvi Konduru <t-konduru@ti.com>