ospi: fix flash test build failure on mpu1_0
status variable is only used in when OSPI_socInit is called, which is
called only for mcu. Declaring this variable for non-mcu cores causing
build to fail as they no longer use this variable. Reusing retVal to get
status of OSPI_socInit instead of having one more status variable only for
OSPI_socInit.

Signed-off-by: Lohith Kumar <l-kumar@ti.com>