[sciclient][SYSFW-7448]: Fix jenkins build issue for DM version feature
Jenkins do the repo init with commit depth equals to 1.
Due to commit depth = 1, git describe command was unable to do reverse
search to find release tags in repository. It leads to jenkins failure.

Added git fetch with depth=500 to device_manager and rm_pm_hal
repositories for which git describe is called. This will fix the issue.

Signed-off-by: Vamshi Vemula <v-vemula@ti.com>